+++ packages/core/src/odata-v4/entity.spec.ts
@@ -7,7 +7,7 @@ import {
   TestEntitySingleLink
 } from '../../test/test-util/test-services/v4/test-service';
 
-describe('entity v4', () => {
+describe('entity', () => {
   let entity;
 
   const state = {
@@ -55,8 +55,8 @@ describe('entity v4', () => {
       })
       .build();
   });
-  it('getCurrentMapKeys', () => {
-    const currentState = entity.getCurrentMapKeys();
+  it('asObject', () => {
+    const currentState = entity.asObject();
 
     expect(currentState).toEqual(state);
     expect(currentState.collectionProperty).not.toBe(entity.collectionProperty);
@@ -70,7 +70,7 @@ describe('entity v4', () => {
     expect(entity.getUpdatedProperties()).toEqual({});
   });
 
-  it('getCurrentMapKeys does not run endlessly', () => {
+  it('asObject() does not run endlessly', () => {
     const parent = TestEntityCircularLinkParent.builder()
       .keyProperty('parent')
       .build();
@@ -86,7 +86,7 @@ describe('entity v4', () => {
     parent.toFirstChild = child1;
     parent.toChildren = [child1, child2];
 
-    expect(parent['getCurrentMapKeys']()).toEqual({
+    expect(parent['asObject']()).toEqual({
       keyProperty: 'parent',
       toFirstChild: {
         keyProperty: 'child1'
@@ -94,7 +94,7 @@ describe('entity v4', () => {
       toChildren: [{ keyProperty: 'child1' }, { keyProperty: 'child2' }]
     });
 
-    expect(child1['getCurrentMapKeys']()).toEqual({
+    expect(child1['asObject']()).toEqual({
       keyProperty: 'child1',
       toParent: {
         keyProperty: 'parent'