+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 
 ## Compatibility Notes
 
--
+- The return type of `Entity.getUpdatedProperties` was incorrected and was corrected to `Record<string, any>`.
 
 ## New Functionality
 
@@ -27,6 +27,8 @@
 ## Fixed Issues
 
 - Fix serialization of linked entities with custom fields.
+- Fix setting of remote state on entities to replace old state instead of merging with old state. This also fixes how custom fields are handled in the remote state.
+- Fix return type of `Entity.getUpdatedProperties` to `Record<string, any>`.
 
 
 # 1.34.0